Move more classes from common to gui
