+++ common/CMakeLists.txt
@@ -23,18 +23,14 @@ target_sources(common PRIVATE
 	PathUtils.cpp
 	PrecompiledHeader.cpp
 	Perf.cpp
-	pxStaticText.cpp
 	pxStreams.cpp
 	pxTranslate.cpp
-	pxWindowTextWriter.cpp
 	RwMutex.cpp
 	Semaphore.cpp
 	SettingsWrapper.cpp
 	StringHelpers.cpp
 	StringUtil.cpp
 	ThreadTools.cpp
-	wxGuiTools.cpp
-	wxHelpers.cpp
 	emitter/bmi.cpp
 	emitter/cpudetect.cpp
 	emitter/fpu.cpp
@@ -74,7 +70,6 @@ target_sources(common PRIVATE
 	PageFaultSource.h
 	PrecompiledHeader.h
 	pxForwardDefs.h
-	pxStaticText.h
 	pxStreams.h
 	RedtapeWindows.h
 	RwMutex.h
@@ -88,7 +83,6 @@ target_sources(common PRIVATE
 	Threading.h
 	TraceLog.h
 	wxBaseTools.h
-	wxGuiTools.h
 	emitter/cpudetect_internal.h
 	emitter/implement/dwshift.h
 	emitter/implement/group1.h
