+++ pcsx2/gui/wxGuiTools.cpp
@@ -13,9 +13,10 @@
  *  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "common/PrecompiledHeader.h"
-#include "common/wxGuiTools.h"
-#include "common/pxStaticText.h"
+#include "PrecompiledHeader.h"
+
+#include "gui/wxGuiTools.h"
+#include "gui/pxStaticText.h"
 
 #include <wx/app.h>
 #include <wx/window.h>
@@ -648,4 +649,34 @@ wxString pxGetAppName()
 {
 	pxAssert(wxTheApp);
 	return wxTheApp->GetAppName();
+}
+
+
+// ------------------------------------------------------------------------
+// Launches the specified file according to its mime type
+//
+void pxLaunch(const wxString& filename)
+{
+	wxLaunchDefaultBrowser(filename);
+}
+
+void pxLaunch(const char* filename)
+{
+	pxLaunch(fromUTF8(filename));
+}
+
+// ------------------------------------------------------------------------
+// Launches a file explorer window on the specified path.  If the given path is not
+// a qualified URI (with a prefix:// ), file:// is automatically prepended.  This
+// bypasses wxWidgets internal filename checking, which can end up launching things
+// through browser more often than desired.
+//
+void pxExplore(const wxString& path)
+{
+	wxLaunchDefaultBrowser(!path.Contains(L"://") ? L"file://" + path : path);
+}
+
+void pxExplore(const char* path)
+{
+	pxExplore(fromUTF8(path));
 }
