+++ pcsx2/gui/wxHelpers.cpp
@@ -13,54 +13,22 @@
  *  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include "PrecompiledHeader.h"
+
 #include <wx/cshelp.h>
 #include <wx/tooltip.h>
 #include <wx/spinctrl.h>
 #include "common/General.h"
-#include "common/wxGuiTools.h"
-#include "common/pxStaticText.h"
 #include "common/Threading.h"
 #include "common/IniInterface.h"
+#include "gui/wxGuiTools.h"
+#include "gui/pxStaticText.h"
 
 using namespace pxSizerFlags;
 
 pxDialogCreationFlags pxDialogFlags()
 {
 	return pxDialogCreationFlags().CloseBox().Caption().Vertical();
-}
-
-
-// --------------------------------------------------------------------------------------
-//  BaseDeletableObject Implementation
-// --------------------------------------------------------------------------------------
-// This code probably deserves a better home.  It's general purpose non-GUI code (the single
-// wxApp/Gui dependency is in wxGuiTools.cpp for now).
-//
-bool BaseDeletableObject::MarkForDeletion()
-{
-	return !m_IsBeingDeleted.exchange(true);
-}
-
-void BaseDeletableObject::DeleteSelf()
-{
-	if (MarkForDeletion())
-		DoDeletion();
-}
-
-BaseDeletableObject::BaseDeletableObject()
-{
-#ifdef _MSC_VER
-// Bleh, this fails because _CrtIsValidHeapPointer calls HeapValidate on the
-// pointer, but the pointer is a virtual base class, so it's not a valid block. >_<
-//pxAssertDev( _CrtIsValidHeapPointer( this ), "BaseDeletableObject types cannot be created on the stack or as temporaries!" );
-#endif
-
-	m_IsBeingDeleted.store(false, std::memory_order_relaxed);
-}
-
-BaseDeletableObject::~BaseDeletableObject()
-{
-	AffinityAssert_AllowFrom_MainUI();
 }
